Extract OAuthErrorAlert into its own component
App.tsx owned the error-code → message map, the dismissed-state
useState, and the Mantine Alert render — all of it OAuth-specific
plumbing inside a top-level container. Move it into
components/OAuthErrorAlert.tsx, which takes `error: string | null`
and owns its own dismissal. App.tsx renders <OAuthErrorAlert
error={oauthError} /> with no surrounding state.
